+++ src/Pencil/Colour.php
@@ -8,6 +8,7 @@
  * PHP version 8.1
  *
  * @package Inane\Cli
+ * @category console
  *
  * @author		Philip Michael Raab<peep@inane.co.za>
  *
@@ -30,12 +31,16 @@ namespace Inane\Cli\Pencil;
  * @version 0.1.0
  */
 enum Colour: int {
-    case Black  = 0;
-    case Red    = 1;
-    case Green  = 2;
-    case Yellow = 3;
-    case Blue   = 4;
-    case Purple = 5;
-    case Cyan   = 6;
-    case White  = 7;
+    /**
+     * Default uses the environments colour
+     */
+    case Default = -1;
+    case Black   = 0;
+    case Red     = 1;
+    case Green   = 2;
+    case Yellow  = 3;
+    case Blue    = 4;
+    case Purple  = 5;
+    case Cyan    = 6;
+    case White   = 7;
 }